requirements.txt
aiogram
environs
redis
betterlogging
yookassa
python-dotenv
gspread
oauth2client
sqlalchemy
asyncpg
alembic
pytz
python-dateutil
